+++ lib/recording_studio_publications/admin.rb
@@ -6,25 +6,35 @@ module RecordingStudioPublications
   module Admin
     SCREEN_KEY = "publications"
     ARTICLES_SCREEN_KEY = "articles"
+    TYPES_SCREEN_KEY = "publication_types"
     SECTION_KEY = "publications"
     RESOURCE_KEY = "publications"
     WIDGET_TOTAL = "widgets.publications.total"
+    WIDGET_ARTICLES_TOTAL = "widgets.articles.total"
     WIDGET_OVER_TIME = "widgets.publications.over_time"
+    WIDGET_ARTICLES_OVER_TIME = "widgets.articles.over_time"
     WIDGET_BY_KIND = "widgets.publications.by_kind"
   end
 end
 
 require_relative "admin/paths"
+require_relative "admin/anchor_urls"
 require_relative "admin/support"
+require_relative "admin/series"
+require_relative "admin/type_rows"
 require_relative "admin/publications_section"
 require_relative "admin/publications_screen"
 require_relative "admin/articles_screen"
+require_relative "admin/publication_types_screen"
 require_relative "admin/publications_resource"
 
 module RecordingStudioPublications
   module Admin
     extend Paths
+    extend AnchorUrls
     extend Support
+    extend Series
+    extend TypeRows
 
     TotalPublicationsWidget = RecordingStudioAdmin::Widget.new(WIDGET_TOTAL, blast_radius: :site) do
       type :number
@@ -35,15 +45,37 @@ module RecordingStudioPublications
       hide_period
     end
 
+    TotalArticlesWidget = RecordingStudioAdmin::Widget.new(WIDGET_ARTICLES_TOTAL, blast_radius: :site) do
+      type :number
+      title I18n.t("recording_studio_publications.admin.articles_total_widget_title", default: "Articles")
+      value { |_context| RecordingStudioPublications.articles.count }
+      link_to { |context| RecordingStudioPublications::Admin.articles_screen_path(context) }
+      hide_change
+      hide_period
+    end
+
     TitlesOverTimeWidget = RecordingStudioAdmin::Widget.new(WIDGET_OVER_TIME, blast_radius: :site) do
       type :chart
-      title I18n.t("recording_studio_publications.admin.over_time_widget_title", default: "Publications over time")
+      title I18n.t("recording_studio_publications.admin.total_widget_title", default: "Publications")
+      value { |_context| RecordingStudioPublications.publications.count }
       chart_type :line
       hide_change
       hide_period
-      hide_metric
       series { |_context| RecordingStudioPublications::Admin.titles_over_time_series }
       chart_options { { height: 220 } }
+      link_to { |context| RecordingStudioPublications::Admin.publications_screen_path(context) }
+    end
+
+    ArticlesOverTimeWidget = RecordingStudioAdmin::Widget.new(WIDGET_ARTICLES_OVER_TIME, blast_radius: :site) do
+      type :chart
+      title I18n.t("recording_studio_publications.admin.articles_total_widget_title", default: "Articles")
+      value { |_context| RecordingStudioPublications.articles.count }
+      chart_type :line
+      hide_change
+      hide_period
+      series { |_context| RecordingStudioPublications::Admin.articles_over_time_series }
+      chart_options { { height: 220 } }
+      link_to { |context| RecordingStudioPublications::Admin.articles_screen_path(context) }
     end
 
     TitlesByKindWidget = RecordingStudioAdmin::Widget.new(WIDGET_BY_KIND, blast_radius: :site) do
@@ -60,12 +92,23 @@ module RecordingStudioPublications
     def self.register!
       return unless defined?(::RecordingStudioAdmin)
 
+      register_screens!
+      register_widgets!
+    end
+
+    def self.register_screens!
       RecordingStudioAdmin.register_section(PublicationsSection)
       RecordingStudioAdmin.register_screen(PublicationsScreen)
       RecordingStudioAdmin.register_screen(ArticlesScreen)
+      RecordingStudioAdmin.register_screen(PublicationTypesScreen)
       RecordingStudioAdmin.register_resource(PublicationsResource)
+    end
+
+    def self.register_widgets!
       RecordingStudioAdmin.register_widget(TotalPublicationsWidget)
+      RecordingStudioAdmin.register_widget(TotalArticlesWidget)
       RecordingStudioAdmin.register_widget(TitlesOverTimeWidget)
+      RecordingStudioAdmin.register_widget(ArticlesOverTimeWidget)
       RecordingStudioAdmin.register_widget(TitlesByKindWidget)
     end
   end